Reopen STDIN/STDOUT/STDERR properly.
For the send_nsca children, not having the /dev/null sink causes a
deadlock / race condition when send_nsca times out talking to a remote
host.  This hangs the main process (which issues a blocking waitpid),
and can cause defunct check processes to stack up.

For check execution, we weren't re-opening STDIN after closing it, which
could be problematic.  Just to keep things clean and consistent, we now
reopen STDIN as </dev/null